in-17.1.18256.18792

+++ BaseApp/Source/Base Application/APIWebhookNotificationMgt.Codeunit.al
@@ -3,6 +3,9 @@ codeunit 6153 "API Webhook Notification Mgt."
     // Registers notifications in table API Webhook Notification on entity insert, modify, rename and delete
 
     SingleInstance = true;
+    Permissions = TableData "API Webhook Subscription" = imd,
+                  TableData "API Webhook Notification" = imd,
+                  TableData "API Webhook Notification Aggr" = imd;
 
     trigger OnRun()
     begin
@@ -35,6 +38,7 @@ codeunit 6153 "API Webhook Notification Mgt."
         UseCachedApiSubscriptionEnabled: Boolean;
         UseCachedDetailedLoggingEnabled: Boolean;
         TooManyJobsMsg: Label 'New job is not created. Count of jobs cannot exceed %1.', Locked = true;
+        NoPermissionsTxt: Label 'No permissions.', Locked = true;
         FieldTok: Label 'Field', Locked = true;
         EqConstTok: Label '=CONST(', Locked = true;
         EqFilterTok: Label '=FILTER(', Locked = true;
@@ -313,6 +317,14 @@ codeunit 6153 "API Webhook Notification Mgt."
     begin
         Session.LogMessage('00006ZQ', StrSubstNo(DeleteSubscriptionMsg,
             DateTimeToString(APIWebhookSubscription."Expiration Date Time"), APIWebhookSubscription."Source Table Id"), Verbosity::Normal, DataClassification::SystemMetadata, TelemetryScope::ExtensionPublisher, 'Category', APIWebhookCategoryLbl);
+
+        if (not APIWebhookSubscription.WritePermission()) or
+           (not APIWebhookNotification.WritePermission()) or
+           (not APIWebhookNotificationAggr.WritePermission()) then begin
+            Session.LogMessage('0000DY1', NoPermissionsTxt, Verbosity::Warning, DataClassification::SystemMetadata, TelemetryScope::ExtensionPublisher, 'Category', APIWebhookCategoryLbl);
+            exit;
+        end;
+
         APIWebhookNotification.SetRange("Subscription ID", APIWebhookSubscription."Subscription Id");
         if not APIWebhookNotification.IsEmpty() then
             APIWebhookNotification.DeleteAll(true);
@@ -327,9 +339,13 @@ codeunit 6153 "API Webhook Notification Mgt."
     local procedure RegisterNotification(var ApiWebhookEntity: Record "Api Webhook Entity"; var APIWebhookSubscription: Record "API Webhook Subscription"; var RecRef: RecordRef; ChangeType: Option): Boolean
     var
         APIWebhookNotification: Record "API Webhook Notification";
-        FieldRef: FieldRef;
         FieldValue: Text;
     begin
+        if not APIWebhookNotification.WritePermission() then begin
+            Session.LogMessage('0000DY2', NoPermissionsTxt, Verbosity::Warning, DataClassification::SystemMetadata, TelemetryScope::ExtensionPublisher, 'Category', APIWebhookCategoryLbl);
+            exit(false);
+        end;
+
         if TryGetEntityKeyValue(ApiWebhookEntity, RecRef, FieldValue) then begin
             APIWebhookNotification.ID := CreateGuid();
             APIWebhookNotification."Subscription ID" := APIWebhookSubscription."Subscription Id";
@@ -338,7 +354,7 @@ codeunit 6153 "API Webhook Notification Mgt."
             if APIWebhookNotification."Change Type" = APIWebhookNotification."Change Type"::Deleted then
                 APIWebhookNotification."Last Modified Date Time" := CurrentDateTime()
             else
-                APIWebhookNotification."Last Modified Date Time" := GetLastModifiedDateTime(RecRef, FieldRef);
+                APIWebhookNotification."Last Modified Date Time" := GetLastModifiedDateTime(RecRef);
             APIWebhookNotification."Entity Key Value" := CopyStr(FieldValue, 1, MaxStrLen(APIWebhookNotification."Entity Key Value"));
             if APIWebhookNotification.Insert(true) then begin
                 if IsDetailedLoggingEnabled() then
@@ -497,18 +513,18 @@ codeunit 6153 "API Webhook Notification Mgt."
             OutBoolean := 'false';
     end;
 
-    local procedure GetLastModifiedDateTime(var RecRef: RecordRef; var FieldRef: FieldRef): DateTime
+    local procedure GetLastModifiedDateTime(var RecRef: RecordRef): DateTime
     var
         LastModifiedDateTime: DateTime;
     begin
-        if FindLastModifiedDateTimeField(RecRef, FieldRef) then
-            LastModifiedDateTime := FieldRef.Value
-        else
+        LastModifiedDateTime := RecRef.Field(RecRef.SystemModifiedAtNo).Value();
+        if LastModifiedDateTime = 0DT then
             LastModifiedDateTime := CurrentDateTime();
         exit(LastModifiedDateTime);
     end;
 
     [Scope('OnPrem')]
+    [Obsolete('Use field SystemModifiedAt', '18.0')]
     procedure FindLastModifiedDateTimeField(var RecRef: RecordRef; var FieldRef: FieldRef): Boolean
     var
         "Field": Record "Field";
